+++ Content.Client/Weapons/Ranged/BulletholeVisualizerSystem.cs
@@ -3,35 +3,21 @@ using Robust.Client.GameObjects;
 
 namespace Content.Client.Weapons.Ranged;
 
-public sealed class BulletholeVisualizerSystem : VisualizerSystem<BulletholeVisualsComponent>
+public sealed class BulletholeVisualizerSystem : VisualizerSystem<BulletholeComponent>
 {
     private const string BulletholeRsiPath = "/Textures/_RMC14/Effects/bulletholes.rsi";
 
-    public override void Initialize()
-    {
-        base.Initialize();
-        SubscribeLocalEvent<BulletholeVisualsComponent, ComponentInit>(OnComponentInit);
-    }
-
-    private void OnComponentInit(EntityUid uid, BulletholeVisualsComponent component, ComponentInit args)
-    {
-        if (!TryComp<SpriteComponent>(uid, out var sprite))
-            return;
-
-    }
-
-    protected override void OnAppearanceChange(EntityUid uid, BulletholeVisualsComponent component, ref AppearanceChangeEvent args)
+    protected override void OnAppearanceChange(EntityUid uid, BulletholeComponent component, ref AppearanceChangeEvent args)
     {
         if (args.Sprite is not { } sprite)
             return;
 
-        if (!AppearanceSystem.TryGetData<string>(uid, BulletholeVisualLayers.State, out var state, args.Component))
+        if (!AppearanceSystem.TryGetData<string>(uid, BulletholeVisuals.State, out var state, args.Component))
             return;
 
         if (!sprite.LayerMapTryGet(BulletholeVisualsLayers.Bullethole, out var layer))
             layer = sprite.LayerMapReserveBlank(BulletholeVisualsLayers.Bullethole);
 
-        Logger.Debug($"Setting sprite to state {state}");
         var valid = !string.IsNullOrWhiteSpace(state);
 
         args.Sprite.LayerSetVisible(BulletholeVisualsLayers.Bullethole, valid);
@@ -44,7 +30,3 @@ public sealed class BulletholeVisualizerSystem : VisualizerSystem<BulletholeVisu
     }
 }
 
-public enum BulletholeVisualsLayers: byte
-{
-    Bullethole
-}
