+++ Content.Shared/_RMC14/Weapons/Ranged/Ammo/BulletholeGeneratorComponent.cs
@@ -1,0 +1,7 @@
+using Robust.Shared.GameStates;
+
+namespace Content.Shared._RMC14.Weapons.Ranged.Ammo;
+
+[RegisterComponent, NetworkedComponent]
+[Access(typeof(BulletholeSystem))]
+public sealed partial class BulletholeGeneratorComponent : Component;
