+++ Resources/Prototypes/_RMC14/Entities/Objects/Weapons/Guns/cm_base_gun.yml
@@ -46,4 +46,4 @@
     layers:
     - state: bullet
       shader: unshaded
-  - type: BallisticAmmo
+  - type: BulletholeGenerator
