+++ Resources/Prototypes/_RMC14/Entities/Structures/Base/base_structure.yml
@@ -126,7 +126,7 @@
   - type: ReceiverXenoClaws
     hitsToDestroy: 20
   - type: Appearance
-  - type: BulletholeVisuals
+  - type: Bullethole
 
 - type: entity
   parent:
